Track1_Patient-Feedback-Reminder/backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
asyncpg
